+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Opendi Solr Client Changelog
 * Added `Core::status()`
 * Added `Client::factory()` for easier client construction
 
+* Moved `Client::ping()` to `Core::ping()`
+* Modified the ping command to take a core name as argument
+
 0.3.2 (2014-08-25)
 ------------------